temp: move `TenderlySimulator` to ./src for external usage on package
